added import/export logic and start refactoring netfront
