+++ front/src/templates/base.html
@@ -299,7 +299,16 @@ height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
 <!-- Main Theme Script -->
 <script src="/assets/js/theme.min.js"></script>
 <script src="{{ url_for('static', filename='/icons.js') }}"></script>
+<script src="{{ url_for('static', filename='/uid-generators.js') }}"></script>
+<script src="{{ url_for('static', filename='/cytoscape-style.js') }}"></script>
+<script src="{{ url_for('static', filename='/config-display.js') }}"></script>
+<script src="{{ url_for('static', filename='/graph-mutations.js') }}"></script>
+<script src="{{ url_for('static', filename='/state-utils.js') }}"></script>
+<script src="{{ url_for('static', filename='/job-editor.js') }}"></script>
+<script src="{{ url_for('static', filename='/grid-canvas.js') }}"></script>
 <script src="{{ url_for('static', filename='/netfront_f.js') }}"></script>
+<script src="{{ url_for('static', filename='/packet-player-ui.js') }}"></script>
+<script src="{{ url_for('static', filename='/api-handlers.js') }}"></script>
 <script src="{{ url_for('static', filename='/miminet_animation.js') }}"></script>
 
 <script>
